wallet: re-fetch all TXOs after importing
After importing, some TXOs that we had previously ignored for not being
ours may now be ours. So we need to re-iterate mapWallet and find those
TXOs.

For the RPCs that unconditionally rescan the entire blockchain if they
are rescanning (importaddress, importpubkey, importprivkey), we don't
need to do this unless they are not rescanning as the rescan will find
the transactions that are already in the wallet and update them
accordingly.