wallet: Change GetBalance to iterate m_txos
Instead of iterating mapWallet and determining the balance by looking at
all of the tx outputs (either directly, or having cached them), use
m_txos and calculate the balance from the stored transaction outputs.

Incidentally, this fixes bug-reorg-spent which happens because of the
balance caches not being updated when a conflicting tx is reorged out.
As the caches are no longer used in the balance calculation, this bug is
fixed.